Add optional configuration imports
This adds the ability to mark imports as `optional`. Optional imports
will be skip loading if they do not exist. If an import is not
explicitly marked as `optional`, then there will be an error raised like
normal.

An optional import is defined as a `path` and `optional` keys. Example:

import:
  - /normal/path/to/config.yml
  - path: /path/to/config.yml
    optional: true

If both these files do not exist alacritty will raise an error on the
/normal/path/to/config.yml file.